+++ src/feed/utilities/index.ts
@@ -1,0 +1,2 @@
+export * from './consume-queue-name';
+export * from './rabbitmq-connection';